# src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ArtHub API",
    "description": "Complete API documentation for ArtHub backend services",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Main API server"
    }
  ],
...
  "paths": {
...
    "/transactions/special-request": {
      "post": {
        "tags": [
          "Transactions"
        ],
        "summary": "Pay for special request",
        "description": "Create a transaction to pay for an accepted special request",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "specialRequestId",
                  "paymentMethod",
                  "shippingAddress"
                ],
                "properties": {
                  "specialRequestId": {
                    "type": "string",
                    "description": "ID of the special request to pay for"
                  },
                  "paymentMethod": {
                    "type": "string",
                    "enum": [
                      "credit_card",
                      "debit_card",
                      "bank_transfer",
                      "paypal",
                      "stripe",
                      "other"
                    ],
                    "description": "Payment method"
                  },
                  "paymentId": {
                    "type": "string",
                    "description": "Payment ID from payment gateway (optional)"
                  },
                  "shippingAddress": {
                    "type": "object",
                    "description": "Shipping address details"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transaction created successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "Special request not found"
          }
        }
      }
    },
...
    "/special-requests": {
      "post": {
        "tags": [
          "Special Requests"
        ],
        "summary": "إنشاء طلب خاص",
        "description": "إرسال طلب خاص لفنان محدد",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "artist",
                  "requestType",
                  "description",
                  "budget"
                ],
                "properties": {
                  "artist": {
                    "type": "string",
                    "description": "معرف الفنان"
                  },
                  "requestType": {
                    "type": "string",
                    "description": "نوع الطلب الخاص"
                  },
                  "description": {
                    "type": "string",
                    "description": "وصف تفصيلي للطلب"
                  },
                  "budget": {
                    "type": "number",
                    "description": "الميزانية المقترحة"
                  },
                  "deadline": {
                    "type": "string",
                    "format": "date-time",
                    "description": "الموعد النهائي (اختياري)"
                  },
                  "attachments": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uri"
                    },
                    "description": "روابط ملفات مرفقة (اختياري)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "تم إنشاء الطلب الخاص بنجاح"
          },
          "400": {
            "description": "بيانات غير صالحة"
          },
          "404": {
            "description": "الفنان غير موجود"
          }
        }
      }
    },
    "/special-requests/my": {
      "get": {
        "tags": [
          "Special Requests"
        ],
        "summary": "عرض طلباتي الخاصة",
        "description": "عرض الطلبات الخاصة التي أرسلها المستخدم",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "رقم الصفحة"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "عدد العناصر في الصفحة"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "accepted",
                "rejected",
                "completed"
              ]
            },
            "description": "تصفية حسب الحالة"
          }
        ],
        "responses": {
          "200": {
            "description": "تم جلب طلبات المستخدم بنجاح"
          }
        }
      }
    },
    "/special-requests/artist": {
      "get": {
        "tags": [
          "Special Requests"
        ],
        "summary": "عرض الطلبات الموجهة للفنان",
        "description": "عرض الطلبات الخاصة المرسلة إلى الفنان",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "رقم الصفحة"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "عدد العناصر في الصفحة"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "accepted",
                "rejected",
                "completed"
              ]
            },
            "description": "تصفية حسب الحالة"
          }
        ],
        "responses": {
          "200": {
            "description": "تم جلب طلبات الفنان بنجاح"
          },
          "403": {
            "description": "غير مصرح لك بعرض طلبات الفنانين"
          }
        }
      }
    },
    "/special-requests/artist/stats": {
      "get": {
        "tags": [
          "Special Requests"
        ],
        "summary": "إحصائيات الطلبات للفنان",
        "description": "عرض إحصائيات الطلبات الخاصة للفنان",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "تم جلب إحصائيات الطلبات بنجاح"
          },
          "403": {
            "description": "غير مصرح لك بعرض إحصائيات الفنانين"
          }
        }
      }
    },
    "/special-requests/{requestId}": {
      "get": {
        "tags": [
          "Special Requests"
        ],
        "summary": "تفاصيل طلب خاص",
        "description": "عرض تفاصيل طلب خاص محدد",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "معرف الطلب"
          }
        ],
        "responses": {
          "200": {
            "description": "تم جلب تفاصيل الطلب بنجاح"
          },
          "404": {
            "description": "الطلب غير موجود"
          }
        }
      }
    },
    "/special-requests/{requestId}/status": {
      "patch": {
        "tags": [
          "Special Requests"
        ],
        "summary": "تحديث حالة الطلب",
        "description": "تحديث حالة طلب خاص (للفنان فقط)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "معرف الطلب"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "accepted",
                      "rejected"
                    ],
                    "description": "الحالة الجديدة"
                  },
                  "response": {
                    "type": "string",
                    "description": "رد الفنان على الطلب"
                  },
                  "deadline": {
                    "type": "string",
                    "format": "date-time",
                    "description": "الموعد النهائي المقترح (في حالة القبول)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "تم تحديث حالة الطلب بنجاح"
          },
          "400": {
            "description": "بيانات غير صالحة"
          },
          "403": {
            "description": "غير مصرح لك بتحديث هذا الطلب"
          },
          "404": {
            "description": "الطلب غير موجود"
          }
        }
      }
    },
    "/special-requests/{requestId}/complete": {
      "patch": {
        "tags": [
          "Special Requests"
        ],
        "summary": "إكمال طلب خاص",
        "description": "تسليم وإكمال طلب خاص (للفنان فقط)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "معرف الطلب"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "deliverables"
                ],
                "properties": {
                  "deliverables": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uri"
                    },
                    "description": "روابط ملفات التسليم"
                  },
                  "finalNote": {
                    "type": "string",
                    "description": "ملاحظة نهائية عند التسليم"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "تم إكمال الطلب بنجاح"
          },
          "400": {
            "description": "بيانات غير صالحة أو الطلب ليس في حالة 'مقبول'"
          },
          "403": {
            "description": "غير مصرح لك بإكمال هذا الطلب"
          },
          "404": {
            "description": "الطلب غير موجود"
          }
        }
      }
    },
...
    "/api/special-requests": {
      "post": {
        "tags": [
          "Special Requests"
        ],
        "summary": "إنشاء طلب خاص",
        "description": "إرسال طلب خاص لفنان محدد",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "artist",
                  "requestType",
                  "description",
                  "budget"
                ],
                "properties": {
                  "artist": {
                    "type": "string",
                    "description": "معرف الفنان"
                  },
                  "requestType": {
                    "type": "string",
                    "description": "نوع الطلب الخاص"
                  },
                  "description": {
                    "type": "string",
                    "description": "وصف تفصيلي للطلب"
                  },
                  "budget": {
                    "type": "number",
                    "description": "الميزانية المقترحة"
                  },
                  "deadline": {
                    "type": "string",
                    "format": "date-time",
                    "description": "الموعد النهائي (اختياري)"
                  },
                  "attachments": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uri"
                    },
                    "description": "روابط ملفات مرفقة (اختياري)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "تم إنشاء الطلب الخاص بنجاح"
          },
          "400": {
            "description": "بيانات غير صالحة"
          },
          "404": {
            "description": "الفنان غير موجود"
          }
        }
      }
    },
    "/api/special-requests/my": {
      "get": {
        "tags": [
          "Special Requests"
        ],
        "summary": "عرض طلباتي الخاصة",
        "description": "عرض الطلبات الخاصة التي أرسلها المستخدم",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "رقم الصفحة"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "عدد العناصر في الصفحة"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "accepted",
                "rejected",
                "completed"
              ]
            },
            "description": "تصفية حسب الحالة"
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "ترتيب حسب الحقل"
          },
          {
            "name": "sortOrder",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            },
            "description": "طريقة الترتيب"
          }
        ],
        "responses": {
          "200": {
            "description": "تم جلب طلبات المستخدم بنجاح"
          }
        }
      }
    },
    "/api/special-requests/artist": {
      "get": {
        "tags": [
          "Special Requests"
        ],
        "summary": "عرض الطلبات الموجهة للفنان",
        "description": "عرض الطلبات الخاصة المرسلة إلى الفنان",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "رقم الصفحة"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "عدد العناصر في الصفحة"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "accepted",
                "rejected",
                "completed"
              ]
            },
            "description": "تصفية حسب الحالة"
          }
        ],
        "responses": {
          "200": {
            "description": "تم جلب طلبات الفنان بنجاح"
          },
          "403": {
            "description": "غير مصرح لك بعرض طلبات الفنانين"
          }
        }
      }
    },
    "/api/special-requests/artist/stats": {
      "get": {
        "tags": [
          "Special Requests"
        ],
        "summary": "إحصائيات الطلبات للفنان",
        "description": "عرض إحصائيات الطلبات الخاصة للفنان",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "تم جلب إحصائيات الطلبات بنجاح"
          },
          "403": {
            "description": "غير مصرح لك بعرض إحصائيات الفنانين"
          }
        }
      }
    },
    "/api/special-requests/{requestId}": {
      "get": {
        "tags": [
          "Special Requests"
        ],
        "summary": "تفاصيل طلب خاص",
        "description": "عرض تفاصيل طلب خاص محدد",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "معرف الطلب"
          }
        ],
        "responses": {
          "200": {
            "description": "تم جلب تفاصيل الطلب بنجاح"
          },
          "404": {
            "description": "الطلب غير موجود"
          }
        }
      },
      "delete": {
        "tags": [
          "Special Requests"
        ],
        "summary": "حذف طلب خاص",
        "description": "حذف طلب خاص (للمرسل فقط وفقط إذا كان الطلب قيد الانتظار أو مرفوضًا)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "معرف الطلب"
          }
        ],
        "responses": {
          "200": {
            "description": "تم حذف الطلب بنجاح"
          },
          "400": {
            "description": "لا يمكن حذف طلب مقبول أو مكتمل"
          },
          "404": {
            "description": "الطلب غير موجود أو غير مصرح بحذفه"
          }
        }
      }
    },
    "/api/special-requests/{requestId}/status": {
      "patch": {
        "tags": [
          "Special Requests"
        ],
        "summary": "تحديث حالة الطلب",
        "description": "تحديث حالة طلب خاص (للفنان فقط)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "معرف الطلب"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "pending",
                      "accepted",
                      "rejected",
                      "completed"
                    ]
                  },
                  "response": {
                    "type": "string",
                    "description": "رد أو ملاحظات إضافية"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "تم تحديث حالة الطلب بنجاح"
          },
          "404": {
            "description": "الطلب غير موجود"
          }
        }
      }
    },
    "/api/special-requests/{requestId}/response": {
      "post": {
        "tags": [
          "Special Requests"
        ],
        "summary": "إضافة رد للطلب",
        "description": "إضافة رد أو تعليق على طلب خاص",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "معرف الطلب"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "response"
                ],
                "properties": {
                  "response": {
                    "type": "string",
                    "description": "الرد أو التعليق"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "تم إضافة الرد بنجاح"
          },
          "404": {
            "description": "الطلب غير موجود"
          }
        }
      }
    },
    "/api/special-requests/{requestId}/complete": {
      "patch": {
        "tags": [
          "Special Requests"
        ],
        "summary": "إكمال الطلب الخاص",
        "description": "تحديث حالة الطلب إلى مكتمل مع إضافة روابط التسليم (للفنان فقط)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "معرف الطلب"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "deliverables": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uri"
                    },
                    "description": "روابط ملفات التسليم"
                  },
                  "finalNote": {
                    "type": "string",
                    "description": "ملاحظة نهائية"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "تم إكمال الطلب بنجاح"
          },
          "400": {
            "description": "لا يمكن إكمال طلب غير مقبول"
          },
          "404": {
            "description": "الطلب غير موجود"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}